Feat: [2025-01-02] 김재연 - 디버깅 연습문제 풀이 2번
