+++ .idea/material_theme_project_new.xml
@@ -1,0 +1,13 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project version="4">
+  <component name="MaterialThemeProjectNewConfig">
+    <option name="metadata">
+      <MTProjectMetadataState>
+        <option name="migrated" value="true" />
+        <option name="pristineConfig" value="false" />
+        <option name="userId" value="4ee9dbe1:193d3ac25c6:-8000" />
+        <option name="version" value="8.4.1" />
+      </MTProjectMetadataState>
+    </option>
+  </component>
+</project>
