+++ src/main/kotlin/YukymController.kt
@@ -5,9 +5,9 @@ import java.time.format.DateTimeFormatter
 
 class YukymController {
 
-    val nowDate = LocalDateTime.now().format(DateTimeFormatter.ofPattern("yyyy-mm-dd"))
+    private val nowDate: String = LocalDateTime.now().format(DateTimeFormatter.ofPattern("yyyy-MM-dd"))
 
-    lateinit var nowTime: String
+    private lateinit var nowTime: String
 
     fun getTyA(): String {
         val timeDataOne = _getTimeDataOne(nowDate)
@@ -32,23 +32,22 @@ class YukymController {
 
     fun getTyB(): String {
         val timeDataOne = _getTimeDataOne(nowDate)
-        var result = timeDataOne.first().ty12
 
         val nowTime = LocalDateTime.now()
-        when {
-            nowTime.hour >= 0 || nowTime.hour < 2 -> return timeDataOne.first().ty1
-            nowTime.hour >= 4 || nowTime.hour < 6 -> return timeDataOne.first().ty2
-            nowTime.hour >= 6 || nowTime.hour < 8 -> return timeDataOne.first().ty3
-            nowTime.hour >= 8 || nowTime.hour < 10 -> return timeDataOne.first().ty4
-            nowTime.hour >= 10 || nowTime.hour < 12 -> return timeDataOne.first().ty5
-            nowTime.hour >= 12 || nowTime.hour < 14 -> return timeDataOne.first().ty6
-            nowTime.hour >= 16 || nowTime.hour < 18 -> return timeDataOne.first().ty7
-            nowTime.hour >= 18 || nowTime.hour < 20 -> return timeDataOne.first().ty8
-            nowTime.hour >= 20 || nowTime.hour < 22 -> return timeDataOne.first().ty9
-            nowTime.hour >= 22 || nowTime.hour < 24 -> return timeDataOne.first().ty10
+        return when (nowTime.hour) {
+            in 0..1 -> timeDataOne.first().ty1
+            in 2..3 -> timeDataOne.first().ty2
+            in 4..5 -> timeDataOne.first().ty3
+            in 6..7 -> timeDataOne.first().ty4
+            in 8..9 -> timeDataOne.first().ty5
+            in 10..11 -> timeDataOne.first().ty6
+            in 12..13 -> timeDataOne.first().ty7
+            in 14..15 -> timeDataOne.first().ty8
+            in 16..17 -> timeDataOne.first().ty9
+            in 18..19 -> timeDataOne.first().ty10
+            in 20..21 -> timeDataOne.first().ty11
+            else -> timeDataOne.first().ty12
         }
-
-        return result
     }
 
     private fun _getTimeDataOne(nowDate: String): List<YukymTimeModel> {
@@ -59,18 +58,3 @@ class YukymController {
         return timeDataOne
     }
 }
-
-data class YukymTimeModel(
-    val ty1: String = "갑자1국",
-    val ty2: String = "갑자2국",
-    val ty3: String = "갑자3국",
-    val ty4: String = "갑자4국",
-    val ty5: String = "갑자5국",
-    val ty6: String = "갑자6국",
-    val ty7: String = "갑자7국",
-    val ty8: String = "갑자8국",
-    val ty9: String = "갑자9국",
-    val ty10: String = "갑자10국",
-    val ty11: String = "갑자11국",
-    val ty12: String = "갑자12국"
-)
