+++ src/main/kotlin/YukymTimeModel.kt
@@ -1,0 +1,16 @@
+package com.survivalcoding
+
+data class YukymTimeModel(
+    val ty1: String = "갑자1국",
+    val ty2: String = "갑자2국",
+    val ty3: String = "갑자3국",
+    val ty4: String = "갑자4국",
+    val ty5: String = "갑자5국",
+    val ty6: String = "갑자6국",
+    val ty7: String = "갑자7국",
+    val ty8: String = "갑자8국",
+    val ty9: String = "갑자9국",
+    val ty10: String = "갑자10국",
+    val ty11: String = "갑자11국",
+    val ty12: String = "갑자12국"
+)
